better interface for time navigation -- fix #17
